Modify(solution) : lab4

+++ solution/lab4/7_6 solution.c
@@ -1,10 +1,10 @@
 int main() {
 
-	printf("Size of int: %d bytes\n", (int)sizeof(int));
-	printf("Size of short: %d bytes\n", (int)sizeof(short));
-	printf("Size of long: %d bytes\n", (int)sizeof(long));
-	printf("Size of float %d bytes\n", (int)sizeof(float));
-	printf("Size of double: %d bytes\n", (int)sizeof(double));
-	printf("Size of long double: %d bytes\n", (int)sizeof(long double));
+	printf("Size of int: %d bytes\n", sizeof(int));
+	printf("Size of short: %d bytes\n", sizeof(short));
+	printf("Size of long: %d bytes\n", sizeof(long));
+	printf("Size of float %d bytes\n", sizeof(float));
+	printf("Size of double: %d bytes\n", sizeof(double));
+	printf("Size of long double: %d bytes\n", sizeof(long double));
 	
 }
